+++ packages/styled-components/src/utils/flatten.js
@@ -14,19 +14,32 @@ import addUnitIfNeeded from './addUnitIfNeeded';
  */
 const isFalsish = chunk => chunk === undefined || chunk === null || chunk === false || chunk === '';
 
-export const objToCss = (obj: Object, prevKey?: string): string => {
-  const css = Object.keys(obj)
-    .filter(key => !isFalsish(obj[key]))
-    .map(key => {
-      if (isPlainObject(obj[key])) return objToCss(obj[key], key);
-      return `${hyphenate(key)}: ${addUnitIfNeeded(key, obj[key])};`;
+
+const objToCssArray = (obj: Object, prevKey?: string): Array<string | Function> => {
+  const rules = []
+  const keys = Object.keys(obj)
+
+  keys
+    .forEach(key => {
+      if (!isFalsish(obj[key])) {
+        if (isPlainObject(obj[key])) {
+          rules.push(...objToCssArray(obj[key], key))
+
+          return rules
+        }
+        else if (isFunction(obj[key])) {
+          rules.push(`${hyphenate(key)}:`, obj[key], ';')
+
+          return rules
+        }
+        rules.push(`${hyphenate(key)}: ${addUnitIfNeeded(key, obj[key])};`);
+      }
+      return rules;
     })
-    .join(' ');
-  return prevKey
-    ? `${prevKey} {
-  ${css}
-}`
-    : css;
+
+    return prevKey
+      ? [`${prevKey} {`, ...rules, '}']
+      : rules;
 };
 
 export default function flatten(chunk: any, executionContext: ?Object, styleSheet: ?Object): any {
@@ -79,5 +92,5 @@ export default function flatten(chunk: any, executionContext: ?Object, styleShee
   }
 
   /* Handle objects */
-  return isPlainObject(chunk) ? objToCss(chunk) : chunk.toString();
+  return isPlainObject(chunk) ? objToCssArray(chunk) : chunk.toString();
 }